+++ src/scope/network/exceptions/ssh-invalid-response.js
@@ -1,0 +1,10 @@
+/** @flow */
+
+export default class SSHInvalidResponse extends Error {
+  response: string;
+
+  constructor(response: string) {
+    super();
+    this.response = response;
+  }
+}